* sprint-july-2:
  Do not send readmarkers for free reads (#1101)
  Display correct size values (#1033)
  Fix exclude path (#1093)
  cleanup preferred blobbers from chain config (#1099)
  Resume download (#1096)
  fix panic (#1090)
  Add support for sending chunks  (#1084)
  fix interface conversion error (#1088)
  Change the contract of Sharder and Miner update settings (#1085)
  Refactor repair (#1004)
  Add path in thumbnail hash (#1013)
  Feature/expose multiop winsdk (#1034)
  Refactor downloads to use sys.File (#1073)
  Multi download operation (#1045)
  feat wasm: readPool unlock (#1076)

+++ core/transaction/entity.go
@@ -323,7 +323,7 @@ func EstimateFee(txn *Transaction, miners []string, reqPercent ...float32) (uint
 					logging.Logger.Error("Object of bad type")
 					return
 				}
-				val := cachedObj.Value.(map[string]interface{})["fee"].(int)
+				val := cachedObj.Value.(map[string]interface{})["fee"].(uint64)
 				feeC <- uint64(val)
 				return
 			}
